pyproject.toml
[project]
name = "discrecontinual-equations"
version = "0.1.0"
description = ""
authors = [
    {name = "John Aven",email = "john.aven.phd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.11.9,<3.0.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "kaleido (>=1.1.0,<2.0.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "sweet-tea (>=0.1.32,<1.0.0)",
    "pytest-cov",
]

[project.optional-dependencies]
dev = ["pytest-cov"]


[tool.pytest.ini_options]
addopts = "--cov=discrecontinual_equations --cov-report=term-missing"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
# Enable pycodestyle (`E`) codes and Pyflakes (`F`) codes by default
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "YTT",  # flake8-2020
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "FBT",  # flake8-boolean-trap
    "A",    # flake8-builtins
    "COM",  # flake8-commas
    "C90",  # McCabe complexity
    "DJ",   # flake8-django
    "EM",   # flake8-errmsg
    "EXE",  # flake8-executable
    "FA",   # flake8-future-annotations
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "G",    # flake8-logging-format
    "INP",  # flake8-no-pep420
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PYI",  # flake8-pyi
    "PT",   # flake8-pytest-style
    "Q",    # flake8-quotes
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SLF",  # flake8-self
    "SLOT", # flake8-slots
    "SIM",  # flake8-simplify
    "TID",  # flake8-tidy-imports
    "TCH",  # flake8-type-checking
    "INT",  # flake8-gettext
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "TD",   # flake8-todos
    "FIX",  # flake8-fixme
    "ERA",  # eradicate
    "PD",   # pandas-vet
    "PGH",  # pygrep-hooks
    "PL",   # pylint
    "TRY",  # tryceratops
    "FLY",  # flynt
    "NPY",  # NumPy-specific rules
    "AIR",  # airflow
    "PERF", # perflint
    "FURB", # refurb
    "LOG",  # flake8-logging
    "RUF",  # ruff-specific rules
]

exclude = [
    "**/__pycache__",
    "**/__pycache__/**",
]

ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Ignore checks for possible passwords
    "S105", "S106", "S107",
    # Allow Greek characters in mathematical docstrings
    "RUF002",
    # Allow imports from typing
    "TCH001", "TCH002", "TCH003",
    # Allow `sys.version` and `sys.version_info`
    "YTT101", "YTT102", "YTT103",
    # Allow `raise NotImplementedError`
    "F901",
]

[tool.ruff.lint.per-file-ignores]
# Tests can use magic values, assertions, and imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
# Examples can use print statements and imports
"examples/**/*" = ["T20", "TID252"]
# Docs can use print statements
"docs/**/*" = ["T20"]

[tool.ruff.lint.isort]
known-first-party = ["discrecontinual_equations"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.flake8-type-checking]
strict = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.0",
    "pymdown-extensions>=10.19.1",
]
